refac: make Transaction a first class concept
